requirements.txt
streamlit==1.39.0
tensorflow==2.16.1
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
Pillow==10.3.0

